feat: moving server configs to be used with `sdk.server()`
